+++ src/main/java/com/twentyone/steachserver/interfaces/status/ServerStatusController.java
@@ -1,4 +1,4 @@
-package com.twentyone.steachserver.util;
+package com.twentyone.steachserver.interfaces.status;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1")
-public class HealthController {
+public class ServerStatusController {
     @GetMapping("/check/server/operating")
     public ResponseEntity<String> CheckServerOperating() {
         return ResponseEntity.ok("OK");
